# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/pullRequests/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Создать PR и автоматически назначить до 2 ревьюверов из команды автора",
                "parameters": [
                    {
                        "description": "pull_request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreatePullRequestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CreatePullRequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequests/merge": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Пометить PR как MERGED (идемпотентная операция)",
                "parameters": [
                    {
                        "description": "pull_request_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MergePullRequestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MergePullRequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequests/reassign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PullRequests"
                ],
                "summary": "Переназначить конкретного ревьювера на другого из его команды",
                "parameters": [
                    {
                        "description": "pull_request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ReassignPullRequestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ReassignPullRequestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teams/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Создать команду с участниками (создаёт/обновляет пользователей)",
                "parameters": [
                    {
                        "description": "team",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/teams/get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Получить команду с участниками",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team_name",
                        "name": "team_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/getReview": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Получить PR'ы, где пользователь назначен ревьювером",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetReviewUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/getStatsHandler": {
            "get": {
                "description": "limit = 0 - убирает лимит",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Получить cтатистику пользователей по PR",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.GetStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/setIsActive": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Установить флаг активности пользователя",
                "parameters": [
                    {
                        "description": "user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SetIsActiveUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SetIsActiveUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AddTeamRequest": {
            "type": "object",
            "required": [
                "members",
                "team_name"
            ],
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string",
                    "example": "payments"
                }
            }
        },
        "model.CreatePullRequestRequest": {
            "type": "object",
            "required": [
                "author_id",
                "pull_request_id",
                "pull_request_name"
            ],
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "u1"
                },
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-1001"
                },
                "pull_request_name": {
                    "type": "string",
                    "example": "Add search"
                }
            }
        },
        "model.CreatePullRequestResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/model.PullRequest"
                }
            }
        },
        "model.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "resource not found"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model.ErrorDetail"
                }
            }
        },
        "model.GetReviewUserResponse": {
            "type": "object",
            "properties": {
                "pull_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PullRequestShort"
                    }
                },
                "user_id": {
                    "type": "string",
                    "example": "u2"
                }
            }
        },
        "model.GetStatsResponse": {
            "type": "object",
            "properties": {
                "user_review_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserStat"
                    }
                }
            }
        },
        "model.MergePullRequestRequest": {
            "type": "object",
            "required": [
                "pull_request_id"
            ],
            "properties": {
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-1001"
                }
            }
        },
        "model.MergePullRequestResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/model.PullRequest"
                }
            }
        },
        "model.PullRequest": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string",
                    "example": "u1"
                },
                "mergedAt": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-1001"
                },
                "pull_request_name": {
                    "type": "string",
                    "example": "Add search"
                },
                "status": {
                    "type": "string",
                    "example": "OPEN"
                }
            }
        },
        "model.PullRequestShort": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string",
                    "example": "u1"
                },
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-1001"
                },
                "pull_request_name": {
                    "type": "string",
                    "example": "Add search"
                },
                "status": {
                    "type": "string",
                    "example": "OPEN"
                }
            }
        },
        "model.ReassignPullRequestRequest": {
            "type": "object",
            "required": [
                "old_reviewer_id",
                "pull_request_id"
            ],
            "properties": {
                "old_reviewer_id": {
                    "type": "string",
                    "example": "u2"
                },
                "pull_request_id": {
                    "type": "string",
                    "example": "pr-1001"
                }
            }
        },
        "model.ReassignPullRequestResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/model.PullRequest"
                },
                "replaced_by": {
                    "type": "string",
                    "example": "u5"
                }
            }
        },
        "model.SetIsActiveUserRequest": {
            "type": "object",
            "required": [
                "is_active",
                "user_id"
            ],
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": false
                },
                "user_id": {
                    "type": "string",
                    "example": "u2"
                }
            }
        },
        "model.SetIsActiveUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "model.Team": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string",
                    "example": "payments"
                }
            }
        },
        "model.TeamMember": {
            "type": "object",
            "required": [
                "is_active",
                "user_id",
                "username"
            ],
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "user_id": {
                    "type": "string",
                    "example": "u1"
                },
                "username": {
                    "type": "string",
                    "example": "Alice"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean",
                    "example": false
                },
                "team_name": {
                    "type": "string",
                    "example": "backend"
                },
                "user_id": {
                    "type": "string",
                    "example": "u2"
                },
                "username": {
                    "type": "string",
                    "example": "Bob"
                }
            }
        },
        "model.UserStat": {
            "type": "object",
            "properties": {
                "active_reviews": {
                    "type": "integer",
                    "example": 1
                },
                "merged_reviews": {
                    "type": "integer",
                    "example": 1
                },
                "total_reviews": {
                    "type": "integer",
                    "example": 2
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string",
                    "example": "u2"
                }
            }
        }
    }
}